# back-end/swagger.json
{
    "swagger": "3.0",
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "AgileOne CRUD",
        "description": "Recruiting Application",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:2000/api/",
            "description": "Dev"
        }
    ],
    "components": {        
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "schemes": ["http","https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/admin/get_clients": {
            "post": {
                "tags": ["Clients"],
                "description": "Get Clients",
                "operationId": "get_clients",
                "summary": "Get Clients Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "perPage": "10/20/30",
                                "page": "Page Number",
                                "search": "Search Text"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/get_allclients": {
            "get": {
                "tags": ["Clients"],
                "description": "Get All Clients",
                "operationId": "get_allclients",
                "summary": "Get All Clients Details in system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/create_client": {
            "post": {
                "tags": ["Clients"],
                "description": "Add Client Details",
                "operationId": "create_client",
                "summary": "Add Client Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "client_name": "Client Name"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/update_client": {
            "post": {
                "tags": ["Clients"],
                "description": "Update Client Details",
                "operationId": "update_client",
                "summary": "Update Client Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "client_name": "Client Name",
                                "id": "Client ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/get_singleclient": {
            "post": {
                "tags": ["Clients"],
                "description": "Get Single Client Details",
                "operationId": "get_singleclient",
                "summary": "Get Single Client Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "user_id": "Client ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}